backend/requirements.txt
# requirements.txt
flask
flask-cors
pillow
torch
torchvision
transformers
openai
numpy
opencv-python
ultralytics
pymongo[srv]
dnspython
python-dotenv